add support for debugging tests
Support for debugging is enabled by modifying the build_spec() and
results() functions in init.lua. Additional functionality is provided
via the dap.lua module.

In order to build the necessary spec for debugging, the name of the
binary that executes the target test or test file must be determined.
This is achieved with dap.get_test_binary().

Since output from the binaries cannot be converted to junit xml, an
auxillary function, dap.translate_results(), is used to convert the
standard rust test output into the result object expected by neotest.

This change will also move the file_exists() function into a dedicated
util.lua module.

This change has been tested with codelldb and lldb-vscode. Other
adapters may or may not require additional updates for compatibility.
Debugging is subject to the same limitations as listed in the README
regarding running tests. Additionally, only the status of the test is
captured when debugging.