Fix running tests for non-default workspace packages
Previously, we would fail to run tests for non-default workspace
packages.

For example, given:

```toml
[workspace]
members = ["path/to/member1", "path/to/member2", "path/to/member3/*"]
default-members = ["path/to/member2", "path/to/member3/foo"]
```

We would not be able to run tests in `path/to/member1`.

This also affected workspaces with a root package, e.g.:

```toml
[workspace]
members = ["bar", "foo"]

[package]
name = "test-ws"
version = "0.1.0"
edition = "2021"
```

We would not be able to run tests in `bar` or `foo`.

Fixes #51.